week5/backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0